+++ client/jetbrains/src/main/java/com/sourcegraph/cody/autocomplete/CodyEditorFactoryListener.java
@@ -17,11 +17,12 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.Disposer;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.sourcegraph.cody.agent.CodyAgent;
-import com.sourcegraph.cody.agent.CodyAgentServer;
+import com.sourcegraph.cody.agent.CodyAgentClient;
 import com.sourcegraph.cody.agent.protocol.Position;
 import com.sourcegraph.cody.agent.protocol.Range;
 import com.sourcegraph.cody.agent.protocol.TextDocument;
 import com.sourcegraph.cody.vscode.InlineAutoCompleteTriggerKind;
+import com.sourcegraph.cody.vscode.InlineCompletionTriggerKind;
 import com.sourcegraph.config.ConfigUtil;
 import java.util.List;
 import org.jetbrains.annotations.NotNull;
@@ -68,7 +69,10 @@ public class CodyEditorFactoryListener implements EditorFactoryListener {
       if (suggestions.isEnabledForEditor(e.getEditor())
           && CodyEditorFactoryListener.isSelectedEditor(e.getEditor())) {
         suggestions.clearAutoCompleteSuggestions(e.getEditor());
-        suggestions.triggerAutoComplete(e.getEditor(), e.getEditor().getCaretModel().getOffset());
+        suggestions.triggerAutoComplete(
+            e.getEditor(),
+            e.getEditor().getCaretModel().getOffset(),
+            InlineCompletionTriggerKind.AUTOMATIC);
       }
     }
   }
@@ -110,7 +114,8 @@ public class CodyEditorFactoryListener implements EditorFactoryListener {
               event.getOldLength() != event.getNewLength()
                   ? InlineAutoCompleteTriggerKind.Invoke
                   : InlineAutoCompleteTriggerKind.Automatic;
-          completions.triggerAutoComplete(this.editor, changeOffset);
+          completions.triggerAutoComplete(
+              this.editor, changeOffset, InlineCompletionTriggerKind.AUTOMATIC);
         }
       }
     }
@@ -177,8 +182,8 @@ public class CodyEditorFactoryListener implements EditorFactoryListener {
     if (editor.getProject() == null) {
       return;
     }
-    CodyAgentServer server = CodyAgent.getServer(editor.getProject());
-    if (server == null) {
+    CodyAgentClient client = CodyAgent.getClient(editor.getProject());
+    if (client.server == null) {
       return;
     }
     VirtualFile file = FileDocumentManager.getInstance().getFile(editor.getDocument());
@@ -190,6 +195,11 @@ public class CodyEditorFactoryListener implements EditorFactoryListener {
             .setFilePath(file.getPath())
             .setContent(editor.getDocument().getText())
             .setSelection(getSelection(editor));
-    server.textDocumentDidChange(document);
+    client.server.textDocumentDidChange(document);
+
+    if (client.codebase == null) {
+      return;
+    }
+    client.codebase.handlePotentialCodebaseChange(editor.getProject(), file);
   }
 }